+++ test/unit/ExceptionHandling/DefaultRequestExceptionHandlerTest.php
@@ -23,39 +23,6 @@ class DefaultRequestExceptionHandlerTest extends \PHPUnit\Framework\TestCase
         $this->assertInstanceOf(DefaultRequestExceptionHandler::class, $this->newSubject());
     }
 
-    public function provider_throwable_types()
-    {
-        $types = [
-            [new \Exception('anything'), TRUE],
-            [new \stdClass, FALSE],
-        ];
-
-        // Only in PHP7
-        if (\class_exists(\Error::class, FALSE)) {
-            $types[] = [new \Error, TRUE];
-        }
-
-        return $types;
-    }
-
-    /**
-     * @dataProvider provider_throwable_types
-     */
-    public function test_it_accepts_throwables_or_throws_invalid_argument($to_handle, $should_accept)
-    {
-        $e = NULL;
-        try {
-            $this->newSubject()->handle($to_handle);
-        } catch (\InvalidArgumentException $e) {
-            // nothing
-        }
-        if ($should_accept) {
-            $this->assertNull($e);
-        } else {
-            $this->assertInstanceOf(\InvalidArgumentException::class, $e);
-        }
-    }
-    
     public function test_it_returns_http_response_from_http_exception()
     {
         $e = \HTTP_Exception::factory(302);